# docs/swagger.yaml
basePath: /
definitions:
...
  user.UserLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  web.WebResponse:
    properties:
      code:
        type: integer
      data: {}
      status:
        type: string
    type: object
info:
  contact:
    name: Tim Pengembang
  description: API Backend Kertas Kerja Kabupaten Madiun
  termsOfService: http://swagger.io/terms/
  title: EKAK Kabupaten Madiun API
  version: "1.0"
paths:
...
  /user/login:
    post:
      consumes:
      - application/json
      description: Autentikasi user untuk mendapatkan JWT Token
      parameters:
      - description: Credentials User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil Login
          schema:
            $ref: '#/definitions/web.WebResponse'
        "400":
          description: Input tidak valid atau Login Gagal
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Login User
      tags:
      - Auth
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"